# Core/contracts/openapi/daemon-api.yaml
openapi: "3.1.0"
info:
  title: ScientificState Local Daemon API
  version: "0.1.0"
  description: |
    Local execution daemon API — localhost only.
    The daemon performs scientific computation; it does NOT assert scientific claims.
    All compute outputs require researcher review before claim evaluation.

    Constitutional boundaries:
    - P7: Non-delegation of scientific authority — daemon executes, human evaluates
    - P1: Raw data immutability — daemon never mutates stored data
    - Localhost boundary: No remote daemon access in M1
  contact:
    name: ScientificState
    url: https://scientificstate.org
  license:
    name: Constitutional Science Infrastructure License
servers:
  - url: "http://127.0.0.1:9473"
    description: Local daemon (Execution_Plan_Phase0.md §4.1 — port 9473)
...
paths:
...
  /domains/{domain_id}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get domain module details
      parameters:
        - name: domain_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain module details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainDetail"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    DomainDetail:
      type: object
      required: [domain_id, domain_name, version, supported_data_types, methods]
      properties:
        domain_id:
          type: string
        domain_name:
          type: string
        version:
          type: string
        supported_data_types:
          type: array
          items:
            type: string
        methods:
          type: array
          items:
            type: object
            required: [method_id, name, produces_uncertainty, produces_validity_scope]
            properties:
              method_id:
                type: string
              name:
                type: string
              produces_uncertainty:
                type: boolean
              produces_validity_scope:
                type: boolean
              compute_class:
                type: string
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        detail:
          type: object